codice/CloudApplication/requirements.txt
gitpython
pyzbar
requests
numpy
imutils
prophet
